heb de auto emails toegevoegd aan de EmailService
